clang: bump clang version
This bumps the hermetic clang from llvmorg-19-init-2941 to
llvmorg-23-init-10931. The bump is needed because the latest MSVC STL
(14.50.35717) used by LUCI Windows bots expands _RAISE / _STL_VERIFY to
__builtin_verbose_trap, a builtin only available in clang 21+. Without
this bump, compiling any TU that includes <string> fails with:

  exception(117,9): error: use of undeclared identifier
  '__builtin_verbose_trap'

The 4-version jump (19 -> 23) also enables several new on-by-default
diagnostics that fire under -Weverything -Werror in the standalone
build. Each one is either advisory, false-positive prone, or scoped
to an ABI/platform that does not apply here, and is added to the
-Wno- list in gn/standalone/BUILD.gn:

  -Wformat-signedness
    Flags signed/unsigned mismatches in printf format specifiers.
    Fires pervasively in PERFETTO_DLOG calls, e.g.
    src/trace_processor/importers/proto/chrome_string_lookup.cc:275
    and src/trace_processor/importers/fuchsia/fuchsia_trace_parser.cc:746.
    The mismatches are benign (debug logging only).

  -Wglobal-constructors
    Clang 23 broadened this warning to flag functions marked with
    __attribute__((constructor)), which are by definition meant to
    run at global init. Fires on InitDebugCrashReporter() in
    src/base/logging.cc:57 and CheckCpuOptimizations() in
    src/base/utils.cc:119 -- both deliberate init hooks.

  -Wlifetime-safety-cross-tu-suggestions
    Advisory suggestion to annotate cross-TU references with
    [[clang::lifetimebound]] (e.g.
    include/perfetto/protozero/scattered_stream_writer.h:67).

  -Wlifetime-safety-dangling-field-moved
    The diagnostic text reads 'This could be a false positive as the
    storage may have been moved later'. Fires across many unit tests
    and importers, e.g.
    src/traced/probes/ps/process_stats_data_source_unittest.cc:112.

  -Wlifetime-safety-dangling-global
    Fires on the intentional pattern of assigning &stack_local to a
    global pointer so a signal handler can reach it, e.g.
    src/tools/dump_ftrace_stats/main.cc:124 (g_task_runner =
    &task_runner) -- the stack frame outlives the global write.

  -Wlifetime-safety-intra-tu-suggestions
    Advisory suggestion to annotate accessors with
    [[clang::lifetimebound]]. Fires e.g. on Status::message() in
    include/perfetto/base/status.h:59 and dozens of other const
    accessors returning a member reference.

  -Wlifetime-safety-invalidation
    False positive on lambdas that capture locals by reference and
    are invoked only within the same function scope, e.g. flush_cpu
    in src/base/cpu_info.cc:83.

  -Wlifetime-safety-return-stack-addr-moved
    Same 'could be false positive' family. Fires e.g. in
    src/trace_processor/importers/common/mapping_tracker.cc:51 and
    src/trace_processor/importers/ftrace/drm_tracker.cc:215.

  -Wlifetime-safety-use-after-scope-moved
    Same family; the diagnostic text itself admits the false-positive
    risk. Fires e.g. in
    src/tracing/ipc/service/relay_ipc_service.cc:66.

  -Wms-bitfield-padding
    Microsoft ABI bit-field packing warning, fired from third-party
    code (buildtools/re2/re2/re2.h:795). Not relevant on Linux/Itanium
    ABI builds.

  -Wnrvo
    Suggests opportunities for NRVO; the emitted code is still
    correct. Fires e.g. in src/base/cpu_info.cc:65 and
    src/base/file_utils.cc:509.

  -Wunique-object-duplication
    Warns about function-local statics being duplicated across
    shared libraries; safe under the C++17 magic-statics rule.
    Fires e.g. on Watchdog::GetInstance() in
    include/perfetto/ext/base/watchdog_noop.h:42.

Note that -Wlifetime-safety is not an umbrella for the
-Wlifetime-safety-* sub-warnings in clang 23 -- each must be listed
independently.

Two real source-level issues surfaced by the upgrade are fixed here:

1. src/trace_processor/core/interpreter/bytecode_interpreter_impl.h
   '~res' (with res of type uint8_t) integer-promotes to int and the
   result was being implicitly narrowed back to a uint8_t store. An
   explicit static_cast<uint8_t> is added to keep
   -Wimplicit-int-conversion quiet without losing the check elsewhere.

2. src/base/http/http_server.cc -- HttpServer::ParseOneWebsocketFrame
   The bytewise unmask loop (payload[i] ^= mask[i % 4]) caused the
   clang 21+ loop vectorizer to emit a load wider than 4 bytes from
   the 4-byte 'mask' array (to splat it into a SIMD register),
   over-reading into the adjacent stack redzone. This trips ASan with
   a stack-buffer-overflow and produces wrong output in optimized
   builds because the over-read bytes are XOR'd in as if they were
   mask bytes. Symptom in HttpServerTest.Websocket:

     expected: 'test message'
     observed: 'test\x94F\xEB\xCAsage'

   (bytes 0..3 and 8..11 unmasked correctly, bytes 4..7 XOR'd with
   stack garbage past mask[3]).

   Rewritten to XOR 4 bytes at a time using an explicit uint32_t
   chunk, which gives the optimizer a correctly-sized mask load.

This makes LUCI work properly on Windows.

+++ src/trace_processor/core/interpreter/bytecode_interpreter_impl.h
@@ -1630,13 +1630,13 @@ inline PERFETTO_ALWAYS_INLINE void CopyToRowLayout(
                                 nbv->bv->count_set_bits_until_in_word(*ptr))
                           : std::numeric_limits<uint32_t>::max();
       uint8_t res = is_non_null ? 0xFF : 0;
-      *dest = invert ? ~res : res;
+      *dest = static_cast<uint8_t>(invert ? ~res : res);
       offset = 1;
     } else if constexpr (std::is_same_v<Nullability, DenseNull>) {
       is_non_null = nbv->bv->is_set(table_index);
       storage_index = table_index;
       uint8_t res = is_non_null ? 0xFF : 0;
-      *dest = invert ? ~res : res;
+      *dest = static_cast<uint8_t>(invert ? ~res : res);
       offset = 1;
     } else {
       static_assert(std::is_same_v<Nullability, NonNull>,
